sniffer/requirements.txt
elasticsearch==5.4.0
argparse==1.2.1
redis==2.10.6
pyserial==3.0.1




